fix chainbridge
