+++ src/indexer/index.ts
@@ -1,47 +1,8 @@
-import {
-  chainbridgeChainid,
-  chainbridgeSendCount,
-  chainbridgeReceiveCount,
-  chainbridgeEvmSendHistory,
-  chainbridgeEvmLimittedSendHistory,
-  chainbridgeEvmRangeSendHistory,
-  chainbridgeEvmReceivedHistory,
-  chainbridgeEvmLimittedReceivedHistory,
-  chainbridgeEvmRangeReceivedHistory,
-  chainbridgeEvmReceiveConfirm,
-} from './chainbridge'
+import {EvmChainBridgeIndexer} from './chainbridge'
 
-import {
-  paraSendCount,
-  paraSendHistory,
-  paraLimittedSendHistory,
-  paraRangeSendHistory,
-  paraReceiveCount,
-  paraReceiveHistory,
-  paraLimittedReceiveHistory,
-  paraRangeReceiveHistory,
-  paraChainbridgeReceiveConfirm,
-} from './para'
+import {ParaIndexer} from './para'
 
 export default {
-  chainbridgeChainid,
-  chainbridgeSendCount,
-  chainbridgeReceiveCount,
-  chainbridgeEvmSendHistory,
-  chainbridgeEvmLimittedSendHistory,
-  chainbridgeEvmRangeSendHistory,
-  chainbridgeEvmReceivedHistory,
-  chainbridgeEvmLimittedReceivedHistory,
-  chainbridgeEvmRangeReceivedHistory,
-  chainbridgeEvmReceiveConfirm,
-
-  paraSendCount,
-  paraSendHistory,
-  paraLimittedSendHistory,
-  paraRangeSendHistory,
-  paraReceiveCount,
-  paraReceiveHistory,
-  paraLimittedReceiveHistory,
-  paraRangeReceiveHistory,
-  paraChainbridgeReceiveConfirm,
+  EvmChainBridgeIndexer,
+  ParaIndexer,
 }
